Implement PR changes from @twicki
