+++ dawn/src/dawn/CodeGen/CodeGen.h
@@ -58,12 +58,6 @@ protected:
 
   void addMplIfdefs(std::vector<std::string>& ppDefines, int mplContainerMaxSize) const;
 
-  bool
-  hasGlobalIndices(const std::shared_ptr<iir::StencilInstantiation>& stencilInstantiation) const;
-  bool hasGlobalIndices(const iir::Stencil& stencil) const;
-
-  void generateGlobalIndices(const iir::Stencil& stencil, Structure& stencilClass) const;
-
   const std::string tmpStorageTypename_ = "tmp_storage_t";
   const std::string tmpMetadataTypename_ = "tmp_meta_data_t";
   const std::string tmpMetadataName_ = "m_tmp_meta_data";
