+++ dawn/test/unit-test/dawn/CodeGen/Naive/CMakeLists.txt
@@ -1,0 +1,19 @@
+##===------------------------------------------------------------------------------*- CMake -*-===##
+##                          _                      
+##                         | |                     
+##                       __| | __ ___      ___ ___  
+##                      / _` |/ _` \ \ /\ / / '_  | 
+##                     | (_| | (_| |\ V  V /| | | |
+##                      \__,_|\__,_| \_/\_/ |_| |_| - Compiler Toolchain
+##
+##
+##  This file is distributed under the MIT License (MIT). 
+##  See LICENSE.txt for details.
+##
+##===------------------------------------------------------------------------------------------===##
+
+set(EXEC_NAME ${PROJECT_NAME}UnittestCodeGenNaive)
+add_executable(${EXEC_NAME} TestCodeGenNaive.cpp)
+target_add_dawn_standard_props(${EXEC_NAME})
+target_link_libraries(${EXEC_NAME} DawnUnittest gtest)
+add_test(NAME ${EXEC_NAME} COMMAND ${EXEC_NAME})
